@@ -1,5 +1,5 @@
 {
-  "name": "docs-oasis-dev",
+  "name": "docs-oasis-io",
   "version": "0.0.0",
   "private": true,
   "scripts": {
@@ -11,29 +11,30 @@
     "deploy": "docusaurus deploy",
     "clear": "docusaurus clear",
     "serve": "docusaurus serve",
-    "test": "jest src",
+    "test": "NODE_OPTIONS='--experimental-vm-modules' jest src",
     "write-translations": "docusaurus write-translations",
     "write-heading-ids": "docusaurus write-heading-ids",
+    "typecheck": "tsc",
     "blc": "linkinator http://localhost:3000 --config linkinator.config.json",
     "lint-markdown": "markdownlint-cli2 'docs/**/*.md' 'docs/**/*.mdx'",
     "fmt-markdown": "markdownlint-cli2-fix 'docs/**/*.md' 'docs/**/*.mdx'"
   },
   "dependencies": {
-    "@docusaurus/core": "^2.4.1",
-    "@docusaurus/plugin-client-redirects": "^2.4.1",
-    "@docusaurus/preset-classic": "^2.4.1",
-    "@docusaurus/remark-plugin-npm2yarn": "^2.4.1",
-    "@easyops-cn/docusaurus-search-local": "^0.34.0",
-    "@mdx-js/react": "^1.6.22",
-    "@mermaid-js/mermaid-cli": "^9.3.0",
-    "clsx": "^1.2.1",
-    "linkinator": "^4.1.2",
-    "markdownlint-cli2": "^0.6.0",
-    "plotly.js-basic-dist": "^2.18.2",
-    "prism-react-renderer": "^1.3.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-plotly.js": "^2.5.1"
+    "@docusaurus/core": "^3.1.0",
+    "@docusaurus/plugin-client-redirects": "^3.1.0",
+    "@docusaurus/preset-classic": "^3.1.0",
+    "@docusaurus/remark-plugin-npm2yarn": "^3.1.0",
+    "@easyops-cn/docusaurus-search-local": "^0.40.1",
+    "@mdx-js/react": "^3.0.0",
+    "@mermaid-js/mermaid-cli": "^10.6.1",
+    "clsx": "^2.0.0",
+    "linkinator": "^6.0.3",
+    "markdownlint-cli2": "^0.11.0",
+    "plotly.js-basic-dist": "^2.27.1",
+    "prism-react-renderer": "^2.3.0",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-plotly.js": "^2.6.0"
   },
   "browserslist": {
     "production": [
@@ -48,8 +49,17 @@
     ]
   },
   "devDependencies": {
-    "@types/plotly.js-basic-dist": "^1.54.0",
-    "@types/react-plotly.js": "^2.5.1",
-    "jest": "^29.0.3"
+    "@docusaurus/module-type-aliases": "^3.1.0",
+    "@docusaurus/tsconfig": "^3.1.0",
+    "@docusaurus/types": "^3.1.0",
+    "@types/plotly.js-basic-dist": "^1.54.4",
+    "@types/react-plotly.js": "^2.6.3",
+    "jest": "^29.7.0",
+    "ts-jest": "^29.1.1",
+    "ts-node": "^10.9.2",
+    "typescript": "~5.2.2"
+  },
+  "engines": {
+    "node": ">=18.0"
   }
 }
