@@ -1,4 +1,6 @@
-const visit = require('unist-util-visit');
+import type {Transformer} from 'unified';
+import type mdast from 'mdast'
+import {visit} from 'unist-util-visit';
 
 const cliRegex = /https:\/\/github\.com\/oasisprotocol\/cli\/blob\/master\/docs\/(.*)\.mdx?(#.*)?/;
 const oasisSdkContractRegex = /https:\/\/github\.com\/oasisprotocol\/oasis-sdk\/blob\/main\/docs\/contract\/(.*)\.mdx?(#.*)?/;
@@ -10,7 +12,7 @@ const docsRegex = /https:\/\/github\.com\/oasisprotocol\/docs\/blob\/main\/docs\
 
 const indexReadmeRegex = /(index|README)($|#)/;
 
-module.exports = function (options) {
+export default function plugin(): Transformer {
     /**
      * Replace github.com URLs pointing to markdown files of the docs/ folders in various github
      * repos with relative website links depending on the specific component.
@@ -23,7 +25,8 @@ module.exports = function (options) {
      * https://github.com/oasisprotocol/docs/blob/main/docs/node/genesis-doc.md#committee-scheduler ->
      *   /node/genesis-doc#committee-scheduler
      */
-    function visitor(node) {
+    function visitor(untypedNode: mdast.Node) {
+        const node = untypedNode as mdast.Definition | mdast.Link
         if (oasisSdkContractRegex.test(node.url)) {
             node.url = node.url.replace(oasisSdkContractRegex, '/dapp/cipher/$1$2');
         } else if (oasisSdkRuntimeRegex.test(node.url)) {
@@ -46,9 +49,7 @@ module.exports = function (options) {
         node.url = node.url.replace(indexReadmeRegex, '$2');
     }
 
-    function transform(tree) {
-        visit(tree, ['definition', 'link', 'linkReference'], visitor);
-    }
-
-    return transform;
+    return (root, _) => {
+        visit(root, ['definition', 'link'], visitor);
+    };
 };
