@@ -1,7 +1,6 @@
-// @ts-check
+import {SidebarsConfig} from '@docusaurus/plugin-content-docs';
 
-/** @type {import('@docusaurus/plugin-content-docs').SidebarsConfig} */
-const sidebars = {
+export const sidebarAdrs: SidebarsConfig = {
   adrs: [
     {
       type: 'category',
@@ -20,5 +19,3 @@ const sidebars = {
     }
   ],
 };
-
-module.exports = sidebars;
