@@ -1,7 +1,6 @@
-// @ts-check
+import {SidebarsConfig} from '@docusaurus/plugin-content-docs';
 
-/** @type {import('@docusaurus/plugin-content-docs').SidebarsConfig} */
-const sidebars = {
+export const sidebarGetInvolved: SidebarsConfig = {
   getInvolved: [
     'get-involved/README',
     {
@@ -48,5 +47,3 @@ const sidebars = {
     'get-involved/token-delivery-and-kyc',
   ],
 };
-
-module.exports = sidebars;
