@@ -1,7 +1,6 @@
-// @ts-check
+import {SidebarsConfig} from '@docusaurus/plugin-content-docs';
 
-/** @type {import('@docusaurus/plugin-content-docs').SidebarsConfig} */
-const sidebars = {
+export const sidebarParaTime: SidebarsConfig = {
   paratime: [
     {
       type: 'doc',
@@ -29,5 +28,3 @@ const sidebars = {
     },
   ],
 };
-
-module.exports = sidebars;
